backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-dotenv==1.0.1
httpx==0.27.0
numpy==1.26.3
reportlab==4.1.0
pydantic==2.7.1 